feat: 뷰홀더를 어댑터로부터 분리
